add utilities and options constants
